Add multer

+++ home/home.services.js
@@ -1,4 +1,6 @@
 const db = require('../_helpers/db');
+const multer = require('multer')
+const fs = require('fs')
 const Home = db.Home
 
 async function postHome(param) {
@@ -6,7 +8,7 @@ async function postHome(param) {
     await home.save()
 }
 
-async function getHome(param) {
+async function getHome() {
     return await Home.find()
 }
 
@@ -24,7 +26,27 @@ async function updateHomeID(id, param) {
 }
 
 async function deleteHomeID(id) {
+    let home = await Home.findById(id)
+
+    if (!home) throw `${id} not found`
+    home.images?.forEach((file) => {
+        fs.unlinkSync(file.path)
+    })
+
     await Home.findByIdAndRemove(id)
+}
+
+function uploadMedias() {
+    const storage = multer.diskStorage({
+      destination: function (req, file, cb) {
+         cb(null, './mediaFiles')
+      },
+      filename: function (req, file, cb) {
+         cb(null, file.originalname);
+      }
+    });
+  
+    return multer({ storage: storage }).array('images', 9)
 }
 
 module.exports = {
@@ -32,5 +54,6 @@ module.exports = {
     getHome,
     getHomeID,
     updateHomeID,
-    deleteHomeID
+    deleteHomeID,
+    uploadMedias
 };
